packages/backend/pyproject.toml
[project]
name = "text2sql"
version = "0.1.0"
description = "Ed-Fi Text-to-SQL backend"
requires-python = ">=3.12"
dependencies = [
    # Core
    "httpx>=0.27",
    "pydantic>=2.7",
    "pyyaml>=6.0",
    "structlog>=24.1",
    "typer>=0.12",
    # Providers
    "openai>=1.50",
    "anthropic>=0.40",
    "numpy>=1.26",
    "faiss-cpu>=1.8",
    "rank-bm25>=0.2.2",
    # SQL
    "sqlalchemy>=2.0",
    "psycopg[binary]>=3.2",
    "pyodbc>=5.1",
    "sqlglot>=25.0",
    "rustworkx>=0.16",
    "msgpack>=1.0",
    "rapidfuzz>=3.9",
    "metaphone>=0.6",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "websockets>=12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.14",
    "ruff>=0.5",
    "mypy>=1.10",
    "types-PyYAML",
]

[project.scripts]
text2sql = "text2sql.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/text2sql"]
